Make the auth login and middleware handler tests real integration tests
