+++ internal/auth/middleware_test.go
@@ -1,34 +1,24 @@
+//go:build integration
+
 package auth_test
 
 import (
-	"errors"
 	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
 
 	. "github.com/starquake/topbanana/internal/auth"
+	"github.com/starquake/topbanana/internal/dbtest"
 	"github.com/starquake/topbanana/internal/session"
+	"github.com/starquake/topbanana/internal/store"
 )
-
-// findCookie returns the first response cookie with the given name and a
-// boolean reporting whether it was found. Used by the EnsurePlayer tests to
-// assert that a fresh session cookie is set on the response.
-func findCookie(rec *httptest.ResponseRecorder, name string) (*http.Cookie, bool) {
-	for _, c := range rec.Result().Cookies() {
-		if c.Name == name {
-			return c, true
-		}
-	}
-
-	return nil, false
-}
 
 func TestRequireGameHost_AllowsAdmin(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	admin, err := store.CreatePlayer(t.Context(), "alice", "alice@example.test", "h", RoleAdmin)
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	admin, err := players.CreatePlayer(t.Context(), "alice", "alice@example.test", "h", RoleAdmin)
 	if err != nil {
 		t.Fatalf("CreatePlayer err = %v, want nil", err)
 	}
@@ -43,7 +33,7 @@ func TestRequireGameHost_AllowsAdmin(t *testing.T) {
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireGameHost(next, store, sessions, nil, discardLogger())
+	mw := RequireGameHost(next, players, sessions, nil, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, admin.ID, 0)
@@ -74,13 +64,21 @@ func TestRequireGameHost_AllowsAdmin(t *testing.T) {
 func TestRequireGameHost_DeniesPlayer(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	// Pre-seed an admin so the next CreatePlayer call is not auto-promoted to admin
-	// by the "first password-bearing registrant becomes admin" rule.
-	if _, err := store.CreatePlayer(t.Context(), "admin", "admin@example.test", "h", RoleAdmin); err != nil {
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	// Pre-seed a credentialled admin so the next CreatePlayer call is not
+	// auto-promoted to admin by the "first password-bearing registrant
+	// becomes admin" rule. (The migration-seeded admin row has no
+	// password_hash, so it does not count for that rule.)
+	if _, err := players.CreatePlayer(
+		t.Context(),
+		"seed-admin",
+		"seed-admin@example.test",
+		"h",
+		RoleAdmin,
+	); err != nil {
 		t.Fatalf("seed admin err = %v, want nil", err)
 	}
-	player, err := store.CreatePlayer(t.Context(), "bob", "bob@example.test", "h", RolePlayer)
+	player, err := players.CreatePlayer(t.Context(), "bob", "bob@example.test", "h", RolePlayer)
 	if err != nil {
 		t.Fatalf("CreatePlayer err = %v, want nil", err)
 	}
@@ -90,7 +88,7 @@ func TestRequireGameHost_DeniesPlayer(t *testing.T) {
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireGameHost(next, store, sessions, nil, discardLogger())
+	mw := RequireGameHost(next, players, sessions, nil, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, player.ID, 0)
@@ -115,12 +113,12 @@ func TestRequireGameHost_DeniesPlayer(t *testing.T) {
 func TestRequireGameHost_NoCookie_RedirectsToLogin(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
 		t.Error("next should not be called without cookie")
 	})
 
-	mw := RequireGameHost(next, store, session.New([]byte("k"), true), nil, discardLogger())
+	mw := RequireGameHost(next, players, session.New([]byte("k"), true), nil, discardLogger())
 
 	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/admin/quizzes", nil)
 	rec := httptest.NewRecorder()
@@ -144,12 +142,12 @@ func TestRequireGameHost_NoCookie_RedirectsToLogin(t *testing.T) {
 func TestRequireGameHost_PostDropsNext(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
 		t.Error("next should not be called without cookie")
 	})
 
-	mw := RequireGameHost(next, store, session.New([]byte("k"), true), nil, discardLogger())
+	mw := RequireGameHost(next, players, session.New([]byte("k"), true), nil, discardLogger())
 
 	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/admin/quizzes", nil)
 	rec := httptest.NewRecorder()
@@ -166,13 +164,13 @@ func TestRequireGameHost_PostDropsNext(t *testing.T) {
 func TestRequireGameHost_UnknownPlayerID_RedirectsToLogin(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
 		t.Error("next should not be called for unknown player")
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireGameHost(next, store, sessions, nil, discardLogger())
+	mw := RequireGameHost(next, players, sessions, nil, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, 9999, 0)
@@ -188,52 +186,29 @@ func TestRequireGameHost_UnknownPlayerID_RedirectsToLogin(t *testing.T) {
 	}
 }
 
-func TestRequireGameHost_StoreError_500(t *testing.T) {
-	t.Parallel()
-
-	store := newStubPlayerStore()
-	admin, err := store.CreatePlayer(t.Context(), "alice", "alice@example.test", "h", RoleAdmin)
-	if err != nil {
-		t.Fatalf("CreatePlayer err = %v, want nil", err)
-	}
-	store.failGet = true
-
-	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
-		t.Error("next should not be called when store errors")
-	})
-
-	sessions := session.New([]byte("k"), true)
-	mw := RequireGameHost(next, store, sessions, nil, discardLogger())
-
-	rec := httptest.NewRecorder()
-	sessions.Set(rec, admin.ID, 0)
-	cookie := rec.Result().Cookies()[0]
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/admin/quizzes", nil)
-	req.AddCookie(cookie)
-	rec = httptest.NewRecorder()
-	mw.ServeHTTP(rec, req)
-
-	if got, want := rec.Code, http.StatusInternalServerError; got != want {
-		t.Errorf("status = %d, want %d", got, want)
-	}
-}
-
 // TestRequireGameHost_AllowsHost pins that a Host (middle tier) reaches the
 // dashboard routes (#538). A pre-seeded admin keeps the new row off the
 // first-registrant promotion; the row is then bumped to Host directly.
 func TestRequireGameHost_AllowsHost(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	if _, err := store.CreatePlayer(t.Context(), "admin", "admin@example.test", "h", RoleAdmin); err != nil {
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	if _, err := players.CreatePlayer(
+		t.Context(),
+		"seed-admin",
+		"seed-admin@example.test",
+		"h",
+		RoleAdmin,
+	); err != nil {
 		t.Fatalf("seed admin err = %v, want nil", err)
 	}
-	host, err := store.CreatePlayer(t.Context(), "hank", "hank@example.test", "h", RolePlayer)
+	host, err := players.CreatePlayer(t.Context(), "hank", "hank@example.test", "h", RolePlayer)
 	if err != nil {
 		t.Fatalf("CreatePlayer err = %v, want nil", err)
 	}
-	host.Role = RoleHost
+	if err = players.SetPlayerRole(t.Context(), host.ID, RoleHost); err != nil {
+		t.Fatalf("SetPlayerRole err = %v, want nil", err)
+	}
 
 	called := false
 	next := http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
@@ -242,7 +217,7 @@ func TestRequireGameHost_AllowsHost(t *testing.T) {
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireGameHost(next, store, sessions, nil, discardLogger())
+	mw := RequireGameHost(next, players, sessions, nil, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, host.ID, 0)
@@ -266,8 +241,8 @@ func TestRequireGameHost_AllowsHost(t *testing.T) {
 func TestRequireAdmin_AllowsAdmin(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	admin, err := store.CreatePlayer(t.Context(), "alice", "alice@example.test", "h", RoleAdmin)
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	admin, err := players.CreatePlayer(t.Context(), "alice", "alice@example.test", "h", RoleAdmin)
 	if err != nil {
 		t.Fatalf("CreatePlayer err = %v, want nil", err)
 	}
@@ -279,7 +254,7 @@ func TestRequireAdmin_AllowsAdmin(t *testing.T) {
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireAdmin(next, store, sessions, discardLogger())
+	mw := RequireAdmin(next, players, sessions, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, admin.ID, 0)
@@ -304,22 +279,30 @@ func TestRequireAdmin_AllowsAdmin(t *testing.T) {
 func TestRequireAdmin_DeniesHostWith404(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	if _, err := store.CreatePlayer(t.Context(), "admin", "admin@example.test", "h", RoleAdmin); err != nil {
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	if _, err := players.CreatePlayer(
+		t.Context(),
+		"seed-admin",
+		"seed-admin@example.test",
+		"h",
+		RoleAdmin,
+	); err != nil {
 		t.Fatalf("seed admin err = %v, want nil", err)
 	}
-	host, err := store.CreatePlayer(t.Context(), "hank", "hank@example.test", "h", RolePlayer)
+	host, err := players.CreatePlayer(t.Context(), "hank", "hank@example.test", "h", RolePlayer)
 	if err != nil {
 		t.Fatalf("CreatePlayer err = %v, want nil", err)
 	}
-	host.Role = RoleHost
+	if err = players.SetPlayerRole(t.Context(), host.ID, RoleHost); err != nil {
+		t.Fatalf("SetPlayerRole err = %v, want nil", err)
+	}
 
 	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
 		t.Error("next should not be called for a host on an admin-only route")
 	})
 
 	sessions := session.New([]byte("k"), true)
-	mw := RequireAdmin(next, store, sessions, discardLogger())
+	mw := RequireAdmin(next, players, sessions, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, host.ID, 0)
@@ -338,7 +321,7 @@ func TestRequireAdmin_DeniesHostWith404(t *testing.T) {
 func TestEnsurePlayer_NoCookie_CreatesAnonymousAndSetsCookie(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	sessions := session.New([]byte("k"), true)
 
 	var seenPlayer *Player
@@ -349,7 +332,7 @@ func TestEnsurePlayer_NoCookie_CreatesAnonymousAndSetsCookie(t *testing.T) {
 		w.WriteHeader(http.StatusTeapot)
 	})
 
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
+	mw := EnsurePlayer(next, players, sessions, discardLogger())
 
 	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
 	rec := httptest.NewRecorder()
@@ -392,8 +375,8 @@ func TestEnsurePlayer_NoCookie_CreatesAnonymousAndSetsCookie(t *testing.T) {
 func TestEnsurePlayer_ValidCookie_ReusesExistingRow(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
-	existing, err := store.CreateAnonymousPlayer(t.Context(), "anon-existing")
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
+	existing, err := players.CreateAnonymousPlayer(t.Context(), "anon-existing")
 	if err != nil {
 		t.Fatalf("CreateAnonymousPlayer err = %v, want nil", err)
 	}
@@ -405,7 +388,7 @@ func TestEnsurePlayer_ValidCookie_ReusesExistingRow(t *testing.T) {
 		seenPlayer, _ = PlayerFromContext(r.Context())
 	})
 
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
+	mw := EnsurePlayer(next, players, sessions, discardLogger())
 
 	rec := httptest.NewRecorder()
 	sessions.Set(rec, existing.ID, 0)
@@ -430,7 +413,7 @@ func TestEnsurePlayer_ValidCookie_ReusesExistingRow(t *testing.T) {
 func TestEnsurePlayer_DeletedPlayer_MintsNewAnonymous(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	sessions := session.New([]byte("k"), true)
 
 	var seenPlayer *Player
@@ -438,7 +421,7 @@ func TestEnsurePlayer_DeletedPlayer_MintsNewAnonymous(t *testing.T) {
 		seenPlayer, _ = PlayerFromContext(r.Context())
 	})
 
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
+	mw := EnsurePlayer(next, players, sessions, discardLogger())
 
 	// Issue a cookie pointing at an ID that does not exist in the store.
 	rec := httptest.NewRecorder()
@@ -467,7 +450,7 @@ func TestEnsurePlayer_DeletedPlayer_MintsNewAnonymous(t *testing.T) {
 func TestEnsurePlayer_TwoCookielessRequests_TwoDistinctPlayers(t *testing.T) {
 	t.Parallel()
 
-	store := newStubPlayerStore()
+	players := store.NewPlayerStore(dbtest.Open(t), discardLogger())
 	sessions := session.New([]byte("k"), true)
 
 	var seenIDs []int64
@@ -476,7 +459,7 @@ func TestEnsurePlayer_TwoCookielessRequests_TwoDistinctPlayers(t *testing.T) {
 		seenIDs = append(seenIDs, p.ID)
 	})
 
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
+	mw := EnsurePlayer(next, players, sessions, discardLogger())
 
 	for range 2 {
 		req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
@@ -488,122 +471,5 @@ func TestEnsurePlayer_TwoCookielessRequests_TwoDistinctPlayers(t *testing.T) {
 	}
 	if seenIDs[0] == seenIDs[1] {
 		t.Errorf("two cookieless requests produced the same player ID %d, want distinct", seenIDs[0])
-	}
-}
-
-func TestEnsurePlayer_PetnameCollision_Retries(t *testing.T) {
-	t.Parallel()
-
-	store := newStubPlayerStore()
-	// Force the first three CreateAnonymousPlayer calls to return
-	// ErrDisplayNameTaken; the fourth attempt should succeed and produce a
-	// regular petname row.
-	store.forceAnonCollisions = 3
-	sessions := session.New([]byte("k"), true)
-
-	var seenPlayer *Player
-	next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-		seenPlayer, _ = PlayerFromContext(r.Context())
-	})
-
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
-	rec := httptest.NewRecorder()
-	mw.ServeHTTP(rec, req)
-
-	if seenPlayer == nil {
-		t.Fatal("PlayerFromContext returned nil player; middleware should have retried past the collisions")
-	}
-	if got := seenPlayer.DisplayName; strings.HasPrefix(got, "anon-") {
-		t.Errorf("seenPlayer.DisplayName = %q, want a petname-style name (no anon- fallback)", got)
-	}
-	if got, want := strings.Count(seenPlayer.DisplayName, "-"), 2; got != want {
-		t.Errorf("seenPlayer.DisplayName = %q, want %d hyphens", seenPlayer.DisplayName, want)
-	}
-}
-
-func TestEnsurePlayer_PetnameExhausted_FallsBackToXid(t *testing.T) {
-	t.Parallel()
-
-	store := newStubPlayerStore()
-	// Force every petname attempt (5) to collide so the fallback xid path
-	// has to run.
-	store.forceAnonCollisions = 5
-	sessions := session.New([]byte("k"), true)
-
-	var seenPlayer *Player
-	next := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-		seenPlayer, _ = PlayerFromContext(r.Context())
-	})
-
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
-	rec := httptest.NewRecorder()
-	mw.ServeHTTP(rec, req)
-
-	if seenPlayer == nil {
-		t.Fatal("PlayerFromContext returned nil player; fallback should have produced one")
-	}
-	if got, want := seenPlayer.DisplayName[:5], "anon-"; got != want {
-		t.Errorf("seenPlayer.DisplayName prefix = %q, want %q (xid fallback after exhausted retries)", got, want)
-	}
-}
-
-func TestEnsurePlayer_CreateAnonymousNonCollisionError_500(t *testing.T) {
-	t.Parallel()
-
-	store := newStubPlayerStore()
-	store.forceAnonErr = errors.New("boom")
-	sessions := session.New([]byte("k"), true)
-
-	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
-		t.Error("next should not be called when CreateAnonymousPlayer returns a non-collision error")
-	})
-
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
-	rec := httptest.NewRecorder()
-	mw.ServeHTTP(rec, req)
-
-	if got, want := rec.Code, http.StatusInternalServerError; got != want {
-		t.Errorf("status = %d, want %d", got, want)
-	}
-}
-
-func TestEnsurePlayer_GetPlayerError_500(t *testing.T) {
-	t.Parallel()
-
-	store := newStubPlayerStore()
-	existing, err := store.CreateAnonymousPlayer(t.Context(), "anon-x")
-	if err != nil {
-		t.Fatalf("CreateAnonymousPlayer err = %v, want nil", err)
-	}
-	store.failGet = true
-
-	sessions := session.New([]byte("k"), true)
-
-	next := http.HandlerFunc(func(_ http.ResponseWriter, _ *http.Request) {
-		t.Error("next should not be called on store error")
-	})
-
-	mw := EnsurePlayer(next, store, sessions, discardLogger())
-
-	rec := httptest.NewRecorder()
-	sessions.Set(rec, existing.ID, 0)
-	cookie := rec.Result().Cookies()[0]
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, "/api/quizzes", nil)
-	req.AddCookie(cookie)
-	rec = httptest.NewRecorder()
-	mw.ServeHTTP(rec, req)
-
-	if got, want := rec.Code, http.StatusInternalServerError; got != want {
-		t.Errorf("status = %d, want %d", got, want)
-	}
-	if got, want := rec.Body.String(), "internal error"; !strings.Contains(got, want) {
-		t.Errorf("body = %q, should contain %q", got, want)
 	}
 }
